requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-rest-framework==0.1.0
djangorestframework==3.8.2
factory-boy==2.11.1
Faker==0.8.16
httpie==0.9.9
idna==2.7
psycopg2==2.7.5
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
six==1.11.0
text-unidecode==1.2
urllib3==1.23
